workshop-module-6/module-6/agentcore-incident-agent/agent_config/memory_hook_provider.py
from bedrock_agentcore.memory import MemoryClient
from strands.hooks.events import AgentInitializedEvent, MessageAddedEvent
from strands.hooks.registry import HookProvider, HookRegistry
import copy

class MemoryHook(HookProvider):
    def __init__(self, memory_client, memory_id, actor_id, session_id):
        self.memory_client = memory_client
        self.memory_id = memory_id
        self.actor_id = actor_id
        self.session_id = session_id

    def on_agent_initialized(self, event):
        try:
            recent_turns = self.memory_client.get_last_k_turns(
                memory_id=self.memory_id, actor_id=self.actor_id,
                session_id=self.session_id, k=5)
            if recent_turns:
                context_messages = []
                for turn in recent_turns:
                    for message in turn:
                        role = "assistant" if message["role"] == "ASSISTANT" else "user"
                        content = message["content"]["text"]
                        context_messages.append({"role": role, "content": [{"text": content}]})
                event.agent.system_prompt += "\nDo not respond with user permissions or operational facts. Use them to know more about the user."
                event.agent.messages = context_messages
        except Exception as e:
            print(f"Memory load error: {e}")

    def _add_context_user_query(self, namespace, query, init_content, event):
        content = None
        memories = self.memory_client.retrieve_memories(
            memory_id=self.memory_id, namespace=namespace, query=query, top_k=3)
        for memory in memories:
            if not content:
                content = "\n\n" + init_content + "\n\n"
            content += memory["content"]["text"]
            if content:
                event.agent.messages[-1]["content"][0]["text"] += content + "\n\n"

    def on_message_added(self, event):
        messages = copy.deepcopy(event.agent.messages)
        try:
            if messages[-1]["role"] in ("user", "assistant"):
                if "text" not in messages[-1]["content"][0]:
                    return
                if messages[-1]["role"] == "user":
                    self._add_context_user_query(
                        namespace=f"incident/{self.actor_id}/context",
                        query=messages[-1]["content"][0]["text"],
                        init_content="These are incident analysis contexts:", event=event)
                    self._add_context_user_query(
                        namespace=f"incident/{self.actor_id}/history",
                        query=messages[-1]["content"][0]["text"],
                        init_content="These are past incident records:", event=event)
                self.memory_client.save_conversation(
                    memory_id=self.memory_id, actor_id=self.actor_id,
                    session_id=self.session_id,
                    messages=[(messages[-1]["content"][0]["text"], messages[-1]["role"])])
        except Exception as e:
            raise RuntimeError(f"Memory save error: {e}")

    def register_hooks(self, registry):
        registry.add_callback(MessageAddedEvent, self.on_message_added)
        registry.add_callback(AgentInitializedEvent, self.on_agent_initialized)
