workshop-module-6/module-6/agentcore-incident-agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
strands-agents-builder>=0.1.0
bedrock-agentcore
mcp>=1.0.0
boto3>=1.35.0
